gamutrflib/pyproject.toml
[tool.poetry]
name = "gamutrflib"
version = "0.5.40"
description = "gamutrf non-gnuradio common code"
authors = ["cglewis <clewis@iqt.org>"]
license = "Apache-2.0"
packages = [
    { include = "gamutrflib" },
]

[tool.poetry.dependencies]
pandas = "2.1.4"
python = ">=3.9,<3.13"
pyzmq = "^25.1.0"
zstandard = "0.22.0"

[tool.poetry.urls]
homepage = "https://github.com/IQTLabs/gamutRF"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
